projects/Store_HL7/EHR_System/requirements.txt
fastapi
uvicorn
requests
hl7apy
streamlit

